requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
Deprecated==1.2.14
ecdsa==0.19.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fastapi==0.115.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
limits==3.13.0
lxml==5.3.0
multidict==6.1.0
numpy==2.0.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pluggy==1.5.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pypdfium2==4.30.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
rsa==4.9
six==1.16.0
slowapi==0.1.9
sniffio==1.3.1
starlette==0.38.6
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
wrapt==1.16.0
yarl==1.13.1
zipp==3.20.2
